Modify getTresspassing to exclude whitelist sites
The first step of the addition of a whitelist feature. If and when a
browser.storage.sync whitelist is created and filled, the urls included
will be excluded from the getTresspassing blacklist. Frontend of loading
and editing a whitelist to be implemented.

The creation of a whitelist is also added here. All that remains is the
frontend of filling it up and editing it.