Make that one can update timeout to a lower time
When the user is out of time, they cannot update the timeout to extend
their time. But they can shorten it if desired, now. So even if one
fails the approveEdit by not having any time left, they can still change
the timeout if the timeout is lower than its previous value.