Release Version 1.3.0
Introduces the major new delay pages feature: Turn it on and set a time, and all
tabs going to a blacklisted page will be delayed for that amount of time before
allowing access to said page, and then starting the actual timer.
Also cleans up code to more efficient and correct versions of methods and
processes, and allows 0 minutes on the timeout, effectively always blocking
blacklisted sites.