Remove the <all_urls> permission
Interestingly, it seems that it isn't even needed. The only case where
privileged functionality is needed is tab.executeScript, and that's
only done on the delay page, which is in Monastery's host and
thus (from FF57 on) automatically allowed.

In background.js, we clarify to only notify us if it's the right
page. This prevents errors about not having permission, since we'll be
told multiple times beforehand about the previous page; saying we only
want ours means that we can't possibly error (hopefully).