Use the native fonts instead of Fira Sans/Open Sans
Fira Sans looks nice, but it isn't native and results in a request to
Google instead of just using the built-in font. The settings panel
will automatically use a good font.

Unfortunately, the popup won't, and will default to `serif`. In
addition, Firefox doesn't support system-ui, so that can't be used
either. Right now, I put it to system-ui, sans-serif so it will use it
when it can, but unfortunately it will be the slightly-uglier
sans-serif right now. I do think that this is better than fetching it
from Google Fonts, though.