Extract the hostname and path from a new valid URL
Fixes #21. Tries to parse the input as a URL. If it fails, then oh
well; otherwise, it will extract the hostname, port and pathname and
put it in automatically. It even recognizes duplicates: if
'google.com' is already in there, then typing 'https://google.com'
will be caught as a duplicate!